StrLen: align implementations
There are two copies of `StrLen` in the RapidJSON code base
 * strfunc.h: rapidjson::internal::StrLen<Ch>
 * unittest.h: Strlen<Ch>

To hide a warning on MSVC, align both implementations to use
'unsigned/SizeType' as return type and add an explicit cast.

+++ test/unittest/unittest.h
@@ -27,10 +27,10 @@
 #endif
 
 template <typename Ch>
-inline size_t StrLen(const Ch* s) {
+inline unsigned StrLen(const Ch* s) {
 	const Ch* p = s;
 	while (*p) p++;
-	return p - s;
+	return unsigned(p - s);
 }
 
 template<typename Ch>
